requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
PyYAML>=6.0
pandas>=1.4.0
numpy>=1.22.0
geopy>=2.2.0
supabase>=0.7.0
python-dateutil>=2.8.2
tabulate>=0.9.0